requirements.txt
python-dotenv
streamlit==1.22.0
numpy>=1.21.0,<2.0.0
pillow>=9.0.0
torch>=2.1.0
torchvision>=0.16.0
tqdm>=4.65.0
opencv-python-headless>=4.7.0,<4.9.0
albumentations>=1.3.0
segmentation_models_pytorch==0.3.3
psutil>=5.9.0
matplotlib>=3.5.0,<3.8.0
google-cloud-storage>=2.10.0
label-studio-sdk>=0.0.25
requests>=2.31.0
boto3>=1.26.0
botocore>=1.29.0
